DjangoAPI/gitsocial/requirements.txt
Django==2.1.5
PyGithub==1.43.5

